some more code

+++ src/api/PickyTenants.WebApi/DataTransferObjects/SearchPropertyDto.cs
@@ -2,8 +2,8 @@ namespace PickyTenants.WebApi.DataTransferObjects;
 
 public class SearchPropertyDto
 {
-    public long Lat { get; set; }
-    public long Lng { get; set; }
+    public double Lat { get; set; }
+    public double Lng { get; set; }
     
     public string Address { get; set; }
